Add files via upload
Computer science was introduced to us in the second year of college. This is a lame project we did in 3 days& nights without prior knowledge of PHP as part of semester project that year.

***, DO NOT USE THIS AS IT IS NOT STANDARD AND it's MOSTLY VULNERABLE ***

I wouldn't like to edit the code as it would take away those night turmoils.
Yeah. The code is lame.
Yes, Vulnerable to SQLi.

Teammate: Juzer Khamosi